+++ src/main/java/roomescape/domain/reservation/ReservationRepository.java
@@ -1,14 +1,16 @@
-package roomescape.repository;
+package roomescape.domain.reservation;
 
 import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import roomescape.domain.reservation.Date;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationStatus;
+import java.util.Optional;
 
-@Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository {
+
+    Reservation save(Reservation reservation);
+
+    List<Reservation> findAll();
+
+    Optional<Reservation> findById(long id);
+
 
     Boolean existsByTimeId(Long id);
 
@@ -29,4 +31,6 @@ public interface ReservationRepository extends JpaRepository<Reservation, Long> 
     Boolean existsByMemberIdAndTimeIdAndThemeIdAndDate(Long memberId, Long timeId, Long themeId, Date date);
 
     List<Reservation> findAllByStatus(ReservationStatus Status);
+
+    Reservation deleteById(long id);
 }
