+++ src/main/java/roomescape/service/ThemeService.java
@@ -4,8 +4,8 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.reservation.ReservationRepository;
 import roomescape.domain.theme.Theme;
-import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
 import roomescape.system.exception.RoomescapeException;
